Refactor Streamlit app into modular architecture
